requirements.txt
paramiko
imageio
pillow
python-dateutil